Error handling improvements, more permissive branch naming, refactors
- Store workflow errors in container to print at end of step
- Add more workflow prep errors
- Rename `abstract_task_id` to `task` for brevity
- allow branch idents to start w/ a number
- refactor plan generation
- reduce node idx size to u16
- moved real values into traverse crate
- interner fns return results instead of panicking
- refactor traversal structs
- new error types, errors for undefined task names

+++ workflow/src/id.rs
@@ -46,9 +46,9 @@ id!(BranchpointId, u8);
 id!(IdentId, u16);
 id!(LiteralId, u8);
 id!(AbstractTaskId, u8);
-id!(AbstractValueId, u8);
+id!(AbstractValueId, u16);
 
-id!(RealTaskId, u8);
-id!(RealValueId, u8);
+id!(RealTaskId, u16);
+id!(RealValueId, u16);
 
 id!(RunStrId, u32);
